Small fix in dim makefile - check dimDid copy via shell expression Syntax fix in run.sh script
git-svn-id: https://subversion.gsi.de/dabc/trunk@878 bcbf6573-9a26-0410-9ebc-ce4ab7aade96

+++ dim/Makefile.mk
@@ -44,12 +44,10 @@ $(DIM_LIB): $(DIMDIR)
 	cd $(DIMDIR); export DIMDIR=$(DIMDIR); export OS=$(DIM_OS); export ODIR=$(DIM_ODIR); export JDK_INCLUDE=$(DIM_JDK_INCLUDE); make -j1 $(DIM_BUILD_ARGS)
 	cp -f $(DIMDIR)/$(DIM_ODIR)/libdim.so $(DIM_LIB)
 ifneq ($(DIM_JDK_INCLUDE),)
-	cp -f $(DIMDIR)/$(DIM_ODIR)/libjdim.so $(DABCDLLPATH)  
+	cp -f $(DIMDIR)/$(DIM_ODIR)/libjdim.so $(DABCDLLPATH)
 endif
-	cp -f $(DIMDIR)/$(DIM_ODIR)/dns $(DABCBINPATH)/dimDns  
-ifneq ($(wildcard $(DIMDIR)/$(DIM_ODIR)/did),)
-	cp -f $(DIMDIR)/$(DIM_ODIR)/did $(DABCBINPATH)/dimDid  
-endif
+	cp -f $(DIMDIR)/$(DIM_ODIR)/dns $(DABCBINPATH)/dimDns
+	if [ -f $(DIMDIR)/$(DIM_ODIR)/did ] ; then cp -f $(DIMDIR)/$(DIM_ODIR)/did $(DABCBINPATH)/dimDid; fi
 	@echo "Dim library build"
 #	cd $(DIMDIR)/jdim/classes; jar cf dim.jar dim/*.class
 #	mv -f $(DIMDIR)/jdim/classes/dim.jar $(DABCDLLPATH)
